backend/requirements.txt
fastapi
uvicorn
oracledb
cryptography
pydantic
python-dotenv
influxdb-client
PyYAML
requests
